+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## v0.0.2 (Unreleased)
+
+- feat: early exit velocity solving for island when converged
+- feat: rename `kcc.setRotation()` to `kcc.setQuaternion()`
+
 ## v0.0.1
 
 - Initial release!